+++ __tests__/__integration__/imperative/__tests__/__integration__/cli/config/init/cli.imperative-test-cli.config.init.integration.test.ts
@@ -40,8 +40,9 @@ describe("imperative-test-cli config init", () => {
             `Initialize config files. Defaults to initializing`,
             `"imperative-test-cli.config.json" in the current working directory unless`,
             `otherwise specified.`,
-            `Use "--user" to init "imperative-test-cli.config.user.json". Use "--global" to`,
-            `initialize the configuration files in your home "~/.zowe" directory.`,
+            `Use "--user-config" to init "imperative-test-cli.config.user.json". Use`,
+            `"--global-config" to initialize the configuration files in your home "~/.zowe"`,
+            `directory.`,
             `Use "--no-prompt" to skip prompting for values in a CI environment.`
         ];
         expectedLines.forEach((line: string) => expect(response.output.toString()).toContain(line));
@@ -63,7 +64,7 @@ describe("imperative-test-cli config init", () => {
     });
     it("should initialize a user project config", () => {
         const response = runCliScript(__dirname + "/__scripts__/init_config.sh",
-            TEST_ENVIRONMENT.workingDir, ["--user --prompt false"]);
+            TEST_ENVIRONMENT.workingDir, ["--user-config --prompt false"]);
         const expectedConfigLocation = path.join(TEST_ENVIRONMENT.workingDir, "test", "imperative-test-cli.config.user.json");
         const expectedSchemaLocation = path.join(TEST_ENVIRONMENT.workingDir, "test", "imperative-test-cli.schema.json");
         expect(response.output.toString()).toContain(`Saved config template to`);
@@ -75,7 +76,7 @@ describe("imperative-test-cli config init", () => {
     });
     it("should initialize a global config", () => {
         const response = runCliScript(__dirname + "/__scripts__/init_config.sh",
-            TEST_ENVIRONMENT.workingDir, ["--global --prompt false"]);
+            TEST_ENVIRONMENT.workingDir, ["--global-config --prompt false"]);
         const expectedConfigLocation = path.join(TEST_ENVIRONMENT.workingDir, "imperative-test-cli.config.json");
         const expectedSchemaLocation = path.join(TEST_ENVIRONMENT.workingDir, "imperative-test-cli.schema.json");
         expect(response.output.toString()).toContain(`Saved config template to`);
@@ -87,7 +88,7 @@ describe("imperative-test-cli config init", () => {
     });
     it("should initialize a user global config", () => {
         const response = runCliScript(__dirname + "/__scripts__/init_config.sh",
-            TEST_ENVIRONMENT.workingDir, ["--global --user --prompt false"]);
+            TEST_ENVIRONMENT.workingDir, ["--global-config --user-config --prompt false"]);
         const expectedConfigLocation = path.join(TEST_ENVIRONMENT.workingDir, "imperative-test-cli.config.user.json");
         const expectedSchemaLocation = path.join(TEST_ENVIRONMENT.workingDir, "imperative-test-cli.schema.json");
         expect(response.output.toString()).toContain(`Saved config template to`);
@@ -112,7 +113,7 @@ describe("imperative-test-cli config init", () => {
     });
     it("should initialize a user project config with prompting", () => {
         const response = runCliScript(__dirname + "/__scripts__/init_config_prompt.sh",
-            TEST_ENVIRONMENT.workingDir, ["--user"]);
+            TEST_ENVIRONMENT.workingDir, ["--user-config"]);
         const expectedConfigLocation = path.join(TEST_ENVIRONMENT.workingDir, "test", "imperative-test-cli.config.user.json");
         const expectedSchemaLocation = path.join(TEST_ENVIRONMENT.workingDir, "test", "imperative-test-cli.schema.json");
         expect(response.output.toString()).not.toContain("Unable to securely save credentials");
@@ -125,7 +126,7 @@ describe("imperative-test-cli config init", () => {
     });
     it("should initialize a global config with prompting", () => {
         const response = runCliScript(__dirname + "/__scripts__/init_config_prompt.sh",
-            TEST_ENVIRONMENT.workingDir, ["--global"]);
+            TEST_ENVIRONMENT.workingDir, ["--global-config"]);
         const expectedConfigLocation = path.join(TEST_ENVIRONMENT.workingDir, "imperative-test-cli.config.json");
         const expectedSchemaLocation = path.join(TEST_ENVIRONMENT.workingDir, "imperative-test-cli.schema.json");
         expect(response.output.toString()).not.toContain("Unable to securely save credentials");
@@ -138,7 +139,7 @@ describe("imperative-test-cli config init", () => {
     });
     it("should initialize a user global config with prompting", () => {
         const response = runCliScript(__dirname + "/__scripts__/init_config_prompt.sh",
-            TEST_ENVIRONMENT.workingDir, ["--global --user"]);
+            TEST_ENVIRONMENT.workingDir, ["--global-config --user-config"]);
         const expectedConfigLocation = path.join(TEST_ENVIRONMENT.workingDir, "imperative-test-cli.config.user.json");
         const expectedSchemaLocation = path.join(TEST_ENVIRONMENT.workingDir, "imperative-test-cli.schema.json");
         expect(response.output.toString()).not.toContain("Unable to securely save credentials");
@@ -167,7 +168,6 @@ describe("imperative-test-cli config init", () => {
     //             }
     //         },
     //         defaults: {},
-    //         plugins: [],
     //         secure: []
     //     };
     //     expect(response.output.toString()).toContain(`Saved config template to`);