+++ packages/config/src/doc/IConfig.ts
@@ -16,6 +16,6 @@ export interface IConfig {
     $schema?: string;
     defaults: { [key: string]: string };
     profiles: { [key: string]: IConfigProfile };
-    plugins: string[];
+    plugins?: string[];
     overrides: IImperativeOverrides;
 }