+++ packages/imperative/__tests__/ConfigurationLoader.test.ts
@@ -11,6 +11,7 @@
 
 import { ConfigurationLoader } from "..";
 import { IImperativeOverrides } from "../src/doc/IImperativeOverrides";
+import { IApimlSvcAttrs } from "../src/doc/IApimlSvcAttrs";
 
 describe("ConfigurationLoader", () => {
   describe("overrides", () => {
@@ -42,5 +43,31 @@ describe("ConfigurationLoader", () => {
 
       expect(result.overrides).toEqual(overrides);
     });
-  });
+
+    it("should return a config with apimlConnLookup", () => {
+        const apimlConnLookup: IApimlSvcAttrs[] = [
+            {
+                apiId: "fake_apiId_1",
+                gatewayUrl: "fake_gatewayUrl_1",
+                connProfType: "fake_connProfType_1"
+            },
+            {
+                apiId: "fake_apiId_2",
+                gatewayUrl: "fake_gatewayUrl_2",
+                connProfType: "fake_connProfType_2"
+            }
+        ]
+
+        const result = ConfigurationLoader.load(
+          {
+            name: "some-name",
+            apimlConnLookup
+          },
+          {},
+          () => undefined
+        );
+
+        expect(result.apimlConnLookup).toEqual(apimlConnLookup);
+      });
+    });
 });