+++ packages/imperative/__tests__/config/cmd/schema/schema.handler.test.ts
@@ -72,8 +72,8 @@ const testProfileConfiguration: IProfileTypeConfiguration[] = [
 ];
 
 const expectedSchemaObj: any = {
-    "$schema": "https://json-schema.org/draft/2019-09/schema#",
-    "$version": 2,
+    "$schema": "https://json-schema.org/draft/2020-12/schema",
+    "$version": 3,
     "type": "object",
     "description": "config",
     "properties": {